% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addClinical.R
\name{addClinical}
\alias{addClinical}
\title{Take a MultiAssayExperiment and include curated variables}
\usage{
addClinical(MultiAssayExperiment, clinicalData)
}
\arguments{
\item{MultiAssayExperiment}{A \linkS4class{MultiAssayExperiment} object}

\item{clinicalData}{A \code{DataFrame} or {data.frame} to merge with
clinical data in the MultiAssayExperiment object}
}
\value{
A \code{\link{MultiAssayExperiment}} object
}
\description{
This function works on the \code{colData} of a
\code{\linkS4class{MultiAssayExperiment}} object to add curated variable
columns. It is recommended that the user run the scripts in the
\code{MultiAssayExperiment-TCGA} repository that build the "enhanced" type
of data. Please see the repository's README for more information.
}
