% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccle.R
\name{ccle}
\alias{ccle}
\title{Set CCLE data parameters}
\usage{
ccle(primary = NULL, standard = TRUE, rangeID = character(),
  idFUN = NULL)
}
\arguments{
\item{primary}{The column name for primary specimen, defaults to
"Tumor_Sample_Barcode" (case ignored).}

\item{standard}{logical (default TRUE) whether to use standard range column
indicators found in ccle data, these include "chromosome",
"start_position", "end_position", and "strand"}

\item{rangeID}{An optional \code{character} vector of length four indicating
genomic positions when standard argument is FALSE}

\item{idFUN}{A function that helps in the parsing of cell line names
(cell line ID extracted by default)}
}
\value{
A \code{list} of arguments for use by \link{makeGRangesList}
}
\description{
This function is for use as an argument to the \link{makeGRangesList}
function for converting raw data. The parameters in this function indicate
properties of CCLE data.
}
\author{
Marcel Ramos \email{mramos09@gmail.com}
}

