% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanExpList.R
\name{cleanExpList}
\alias{cleanExpList}
\title{Remove unmatched samples from a list of experiments}
\usage{
cleanExpList(exlist, mPheno)
}
\arguments{
\item{exlist}{A named \linkS4class{Elist} of experiments compatible with the
MultiAssayExperiment API}

\item{mPheno}{A \code{data.frame} of clinical data with patient identifiers
as rownames}
}
\value{
A named \code{list} of experiments
}
\description{
This function is intended to drop any unmatched samples from all of the
listed experiments that are not present in the rownames of the pheno data.
}
\author{
Marcel Ramos \email{mramos09@gmail.com}
}

