% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clinicalNames}
\alias{clinicalNames}
\title{Clinical dataset names in TCGA}
\format{
A \link[IRanges:CharacterList-class]{CharacterList} of names for 33
cancer codes
}
\usage{
data("clinicalNames")
}
\value{
The clinical dataset column names in TCGA as provided by the
\code{RTCGAToolbox}
}
\description{
A dataset of names for each of the TCGA cancer codes available.
These names were obtained by the clinical datasets from
\link[RTCGAToolbox:getFirehoseData]{getFirehoseData}. They serve to subset the
current datasets provided by \code{curatedTCGAData}.
}
\keyword{datasets}
