% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAenhanceClinical.R
\name{enhanceClinical}
\alias{enhanceClinical}
\title{Take a MultiAssayExperiment and include curated variables}
\usage{
enhanceClinical(MultiAssayExperiment = MultiAssayExperiment(), cancerCode,
  pkgLocation = ".")
}
\arguments{
\item{MultiAssayExperiment}{}
}
\value{
A \code{\link{MultiAssayExperiment}} object
}
\description{
This function works on the \code{pData} of a
\code{\link{MultiAssayExperiment}} object to add curated variable columns.
Endomorphic operation!
}

