% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract}
\alias{extract}
\title{Extract data from \code{FirehoseData} object into \code{ExpressionSet} or
\code{GRangesList} object}
\usage{
extract(object, type = NULL)
}
\arguments{
\item{object}{A \code{FirehoseData} object from which to extract data.}

\item{type}{The type of data to extract from the "FirehoseData" object,
see type section.}
}
\value{
Either an \linkS4class{ExpressionSet} object or a
\linkS4class{GRangesList} object.
}
\description{
This function obtains and processes data from a
\link[RTCGAToolbox]{FirehoseData} object. Processing involves converting
raw data to structured S4 Bioconductor objects. The function returns an
\linkS4class{ExpressionSet} or \linkS4class{GRangesList} class object.
}
\section{type}{

Choices include: "RNAseq_Gene",
"Clinic", "miRNASeq_Gene", "RNAseq2_Gene_Norm", "CNA_SNP", "CNV_SNP",
"CNA_Seq", "CNA_CGH", "Methylation", "Mutation", "mRNA_Array",
"miRNA_Array", "RPPA_Array", "GISTIC_A", "GISTIC_T". The "GISTIC_A" type of
dataset represents GISTIC data by all genes. "GISTIC_T" represents data
thresholded by genes. Lowercase entries and entries without the "underscore"
character are also valid inputs.
}
\examples{

\dontrun{
b2 <- extract(a2, "Methylation")
}

}
\author{
Marcel Ramos \email{mramos09@gmail.com}
}

