% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findGRangesCols.R
\name{findGRangesCols}
\alias{findGRangesCols}
\title{Obtain minimum necessary names for the creation of a GRangesList object}
\usage{
findGRangesCols(df_colnames, seqnames.field = c("seqnames", "seqname",
  "chromosome", "chrom", "chr", "chromosome_name", "seqid"),
  start.field = "start", end.field = c("end", "stop"),
  strand.field = "strand", ignore.strand = FALSE)
}
\arguments{
\item{namesVector}{A \code{character} vector or names in a dataset}
}
\value{
Index positions vector indicating columns with appropriate names
}
\description{
This function attempts to match chromosome, start position, end position and
strand names in the given character vector.
}
\examples{
myDataColNames <- c("Start_position", "End_position", "strand",
                 "chromo", "num_probes", "segment_mean")
findGRangesCols(myDataColNames)

}

