% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateMap.R
\name{generateMap}
\alias{generateMap}
\title{Create a sampleMap from an experiment list and phenoData dataframe}
\usage{
generateMap(experiments, colData, idConverter = NULL, ...)
}
\arguments{
\item{experiments}{A named \code{list} of experiments compatible with the
MultiAssayExperiment API}

\item{colData}{A \code{data.frame} of clinical data with patient identifiers
as rownames}

\item{idConverter}{A function to be used against the sample or specimen
identifiers to match those in the rownames of the \code{colData}
(default NULL)}

\item{...}{Additonal arguments to pass to the 'idConverter' function.}
}
\value{
A \code{DataFrame} class object of mapped samples and patient
identifiers including assays
}
\description{
This function helps create a sampleMap in preparation of a
\code{MultiAssayExperiment} object. This especially useful when the
sample identifiers are not very different, as in the case of TCGA barcodes.
An \code{idConverter} function can be provided to truncate such sample
identifiers and obtain patient identifiers.
}
\examples{
\dontrun{
## For TCGA data
newMap <- generateMap(myExpList, myPheno, TCGAbarcode)
}

}
\author{
Marcel Ramos, Martin Morgan, Lucas Schiffer
}
