% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateTCGAmap.R
\name{generateTCGAmap}
\alias{generateTCGAmap}
\title{Create a sampleMap from an experiment list and phenoData dataframe}
\usage{
generateTCGAmap(exlist, mPheno)
}
\arguments{
\item{exlist}{A named \code{list} of experiments compatible with the
MultiAssayExperiment API}

\item{mPheno}{A \code{data.frame} of clinical data with patient identifiers
as rownames}
}
\value{
A \code{DataFrame} class object of mapped samples and patient
identifiers including assays
}
\description{
This function helps create a sampleMap in preparation of a
\code{MultiAssayExperiment} object
}
\author{
Marcel Ramos \email{mramos09@gmail.com}
}

