% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFileNames.R
\name{getFileNames}
\alias{getFileNames}
\title{Find the file names used in RTCGAToolbox}
\usage{
getFileNames(disease, runDate = "20160128", CNASNP = FALSE,
  CNVSNP = FALSE, CNASeq = FALSE, CNACGH = FALSE)
}
\arguments{
\item{disease}{The TCGA cancer disease code, e.g., "COAD"}

\item{runDate}{The single \code{string} used in the \code{getFirehoseData}
function (default "20160128")}

\item{CNASNP}{A \code{logical} (default = FALSE) vector indicating whether
to get the file name from this data type}

\item{CNVSNP}{A \code{logical} (default = FALSE) vector indicating whether
to get the file name from this data type}

\item{CNASeq}{A \code{logical} (default = FALSE) vector indicating whether
to get the file name from this data type}

\item{CNACGH}{A \code{logical} (default = FALSE) vector indicating whether
to get the file name from this data type}
}
\value{
A \code{character} vector of length one indicating the file name
}
\description{
Part of this function is from the RTCGAToolbox. It aims to extract the file
name used inside of the \link[RTCGAToolbox]{getFirehoseData} function.
The arguments of the function parallel those in the
\link[RTCGAToolbox]{getFirehoseData} function. It is only available for
select data types.
}
\examples{

getFileNames("COAD", CNVSNP = TRUE)

}
