% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGRangesList.R
\name{makeGRangesList}
\alias{makeGRangesList}
\title{Convert raw Mutation data to GRangesList}
\usage{
makeGRangesList(inputData, dataparam = NULL, ...)
}
\arguments{
\item{inputData}{A \code{data.frame} or \code{list} class of mutation data}

\item{dataparam}{A function for specifying a list of parameters to pass to
the function depending on the type of data}

\item{...}{Additional arguments passed to the identifier parser function}
}
\value{
A \linkS4class{GRangesList} class object
}
\description{
This function takes the data.frame or list of raw data and converts it to
a \linkS4class{GRangesList} class. Input data can be entered along with a
function that indicates the parameters of the data such as, range column
names, sample/specimen identifier column name and an ID parsing function
(such as TCGAbarcodes). If input data is entered as a list, all
element names are expected to be sample/specimen identifiers. See data
specific functions for more details.
}
\examples{
\dontrun{
## TCGA example
library(TCGAbiolinks)
meso <- TCGAquery_maf(tumor = "meso")
## Select number 2
makeGRangesList(meso, tcga(primary = "Tumor_Sample_Barcode",
                             standard = TRUE)
         sample = TRUE, collapse = TRUE)
}

}
\author{
Marcel Ramos \email{mramos09@gmail.com}
}
\seealso{
tcga(), ccle()
}

