% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGRangesListFromTCGA.R
\name{makeGRangesListFromTCGA}
\alias{makeGRangesListFromTCGA}
\title{Make a GRangesList from TCGA data}
\usage{
makeGRangesListFromTCGA(x, primary, feature.field = "hugo_symbol", ...)
}
\arguments{
\item{x}{A \code{data.frame} or \code{DataFrame} class object. \code{list}
class objects are coerced to data.frame or DataFrame.}

\item{primary}{A \code{character} vector of length one indicating the
column to be used as sample identifiers}

\item{feature.field}{A \code{character} vector of length one indicating the
column to be used as names for each of the ranges in the data}

\item{...}{Additional arguments to pass on to
\link{makeGRangesListFromDataFrame}}
}
\value{
A \link{GRangesList} class object
}
\description{
\code{makeGRangesListFromTCGA} allows the user to convert objects of class
data.frame or DataFrame to a \link{GRangesList}. It includes additional
features specific to TCGA data such as, hugo symbols, probe numbers,
segment means, and ucsc build (if available).
}

