% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeRSE.R
\name{makeRSE}
\alias{makeRSE}
\title{Create a RangedSummarizedExperiment from a data frame}
\usage{
makeRSE(inputData, geneCol = character(), regEx)
}
\arguments{
\item{inputData}{A \code{data.frame} of ranged and expression data}

\item{geneCol}{The name of the gene symbol column as character}

\item{regEx}{A vector of regular expressions for "grepping" each of the 4
columns needed to create the necessary \link{GRanges} object}
}
\value{
A \link{SummarizedExperiment} object with rowRanges
}
\description{
This function uses input data and some regular expression to guess or
indicate the method for matching ranged column names. Such column names are
used to create \code{rowRanges} for the \link{SummarizedExperiment}.
}
\author{
Marcel Ramos \email{mramos09@gmail.com}
}

