% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSummarizedExperimentFromGISTIC.R
\name{makeSummarizedExperimentFromGISTIC}
\alias{makeSummarizedExperimentFromGISTIC}
\title{Create a SummarizedExperiment from FireHose GISTIC}
\usage{
makeSummarizedExperimentFromGISTIC(gistic, dataType)
}
\arguments{
\item{gistic}{A \linkS4class{FirehoseGISTIC} object}

\item{dataType}{Either one of "ThresholdedByGene", "AllByGene", "Peaks"}
}
\description{
Use the output of \code{getFirehoseData} to create a
\linkS4class{SummarizedExperiment}. This can be done for three types of
data, G-scores thresholded by gene, copy number by gene, and copy number by
peak regions.
}
\examples{
library(RTCGAToolbox)
co <- getFirehoseData("COAD", clinical = FALSE, GISTIC = TRUE)
makeSummarizedExperimentFromGISTIC(co, "AllByGene")

}
\author{
Ludwig Geistlinger, Marcel Ramos
}
