% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchClinical.R
\name{matchClinical}
\alias{matchClinical}
\title{Match Clinical with Experiment Data}
\usage{
matchClinical(expData, phenoDat)
}
\arguments{
\item{expData}{Either a \code{list} or \code{data.frame} of data where
identifiers are element names or column names, respectively}

\item{phenoDat}{A \code{data.frame} of clinical variables where rownames
indicate patient identifiers to match on}
}
\value{
A corresponding \code{data.frame} of clinical variables of the same
size as the number of unique patient identifiers with patient identifiers as
a "patientids" column
}
\description{
This function parses patient barcode identifiers to find matches in both
the experiment data and the clinical datasets. It creates a corresponding
clinical dataset for all of the unique patients in the experiment dataset.
}
\author{
Marcel Ramos \email{mramos09@gmail.com}
}

