% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplifyTCGA.R
\name{qreduceTCGA}
\alias{qreduceTCGA}
\title{Simplify curatedTCGAData objects by replacing RaggedExperiment objects}
\usage{
qreduceTCGA(obj, keep = FALSE, suffix = "_simplified")
}
\arguments{
\item{obj}{A MultiAssayExperiment object obtained from curatedTCGAData}

\item{keep}{logical (default FALSE) whether to remove the original
RaggedExperiment objects from the returned MultiAssayExperiment}

\item{suffix}{character (default "_simplified") A character string to append
to the newly modified assay.}
}
\value{
A MultiAssayExperiment object with RaggedExperiments converted to
RangedSummarizedExperiment with rows corresponding to gene symbol.

"Mutations" objects become a genes x patients RangedSummarizedExperiment
containing 1 if there is a non-silent mutation somewhere in the gene, and 0
otherwise. "CNA" and "CNV" segmented copy number are reduced using a weighted
mean in the rare cases of overlapping (non-disjoint) copy number regions.
}
\description{
Simplify curatedTCGAData objects by replacing RaggedExperiment objects
}
\details{
Relies on TxDb.Hsapiens.UCSC.hg19.knownGene and org.Hs.eg.db to map to hg19
NCBI build.
}
\examples{
library(curatedTCGAData)

accmae <-
    curatedTCGAData("ACC", c("CNASNP", "Mutation"), dry.run = FALSE)

qreduceTCGA(accmae)
}
\author{
L. Waldron
}
