% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readExonFiles.R
\name{readExonFiles}
\alias{readExonFiles}
\title{Read Exon level files and create a GRangesList}
\usage{
readExonFiles(filepaths)
}
\arguments{
\item{filepaths}{A vector of valid exon data file paths}
}
\value{
A \linkS4class{GRangesList} object
}
\description{
This function serves to read exon level data from a vector of file paths
and to create a \linkS4class{GRangesList} object for representation. File
name and structure requirements are as follows: The third atomic position
of the parsed file name should be the universally unique identifier (UUID).
File should use default separators (i.e., periods). The column containing
ranged information should be named "exon."
}
\author{
Marcel Ramos \email{mramos09@gmail.com}
}

