% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampleTypes}
\alias{sampleTypes}
\title{Barcode Sample Type Table}
\format{A data frame with 19 rows and 3 variables:
\describe{
     \item{Code}{Two digit code number found in the barcode}
     \item{Definition}{Long name for the sample type}
     \item{Short.Letter.Code}{Letter code for the sample type}
}}
\source{
\url{https://gdc.cancer.gov/resources-tcga-users/tcga-code-tables/sample-type-codes}
}
\usage{
sampleTypes
}
\value{
The TCGA `sampleTypes` table
}
\description{
A dataset that contains the mappings for sample codes in the TCGA
barcodes.
}
\keyword{datasets}
