% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplifyTCGA.R
\name{simplifyTCGA-defunct}
\alias{simplifyTCGA-defunct}
\alias{mirToRanges}
\title{Defunct TCGAutils functions}
\usage{
mirToRanges(obj, keep.assay = FALSE, unmapped = TRUE)
}
\arguments{
\item{obj}{A \code{MultiAssayExperiment} object obtained from \code{curatedTCGAData}}

\item{keep.assay}{logical (default FALSE) Whether to keep the
\code{SummarizedExperiment} assays that have been converted to
\code{RangedSummarizedExperiment}}

\item{unmapped}{logical (default TRUE) Include an assay of data that was
not able to be mapped in reference database}
}
\description{
\code{mirToRanges} is defunct and will be removed in the next
release. The \code{mirbase.db} package is currently deprecated in \code{RELEASE_3_21}.
}
