% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga.R
\name{tcga}
\alias{tcga}
\title{Set TCGA data parameters}
\usage{
tcga(standard = TRUE, single = FALSE, targetCol = character(),
  idFUN = NULL)
}
\arguments{
\item{standard}{logical (default TRUE) whether to use standard range column
indicators found in the tcga data, these include "chromosome",
"start_position", "end_position", and "strand".}

\item{single}{logical (default FALSE) whether to use a single column for
the conversion of ranged values. Such column may look like: "chr1:100-200:+"}

\item{targetCol}{A \code{character} string that indicates the name of the
column to use when a single column is used (i.e., TRUE)}

\item{idFUN}{A function that helps in the parsing of barcode identifiers
(barcode used by default)}
}
\value{
A \code{list} of arguments for use by \link{makeGRangesList}
}
\description{
This function is for use inside the \link{makeGRangesList} function for
converting raw data. The parameters in this function indicate properties
of the TCGA data.
}
\author{
Marcel Ramos \email{mramos09@gmail.com}
}

