% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAbarcode.R
\name{TCGAbarcode}
\alias{TCGAbarcode}
\title{Parse data from TCGA barcode}
\usage{
TCGAbarcode(barcodes, position = NULL, part = TRUE, sample = FALSE,
  vial = FALSE, portion = FALSE, plate = FALSE, center = FALSE,
  collapse = FALSE)
}
\arguments{
\item{barcodes}{A character vector of TCGA barcodes}

\item{position}{A numerical vector of TCGA barcode positions desired}

\item{part}{Logical (default TRUE) participant identifier chunk}

\item{sample}{Logical (default FALSE) includes the numeric sample code of
the barcode}

\item{vial}{Logical (default FALSE) includes the sample vial label}

\item{portion}{Logical (default FALSE) includes the portion and analyte
codes of the barcode}

\item{plate}{Logical (default FALSE) returns the plate value}

\item{center}{Logical (default FALSE) returns a matrix with the plate and
center codes}

\item{collapse}{Logical (default FALSE) concatenates requested barcode
chunks}
}
\value{
A character vector or data matrix of TCGA barcode information
}
\description{
This function returns the specified snippet of information obtained from
the TCGA barcode.
}
\examples{
library(TCGAbiolinks)
query <- TCGAquery(tumor = "OV", level = 3)
barcodes <- unlist(strsplit(query$barcode[1], ","))
TCGAsamplebc <- TCGAbarcode(barcodes, sample = TRUE, collapse = TRUE)

}
\author{
Marcel Ramos \email{mramos09@gmail.com}
}

