% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAbiospec.R
\name{TCGAbiospec}
\alias{TCGAbiospec}
\title{Extract biospecimen data from the TCGA barcode}
\usage{
TCGAbiospec(barcodes)
}
\arguments{
\item{barcodes}{A character vector of TCGA barcodes}
}
\value{
A \code{dataframe} with sample type, sample code, portion, plate,
and center columns.
}
\description{
This function uses the full TCGA barcode to return a data frame of the
data pertinent to laboratory variables such as vials, portions, analytes,
plates and the center.
}
\examples{
example("TCGAbarcode")
right_bctable <- TCGAbiospec(barcodes)

}
\author{
Marcel Ramos \email{mramos09@gmail.com}
}

