% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAcleanExpList.R
\name{TCGAcleanExpList}
\alias{TCGAcleanExpList}
\title{Remove unmatched samples from a list of experiments}
\usage{
TCGAcleanExpList(exlist, mPheno)
}
\arguments{
\item{exlist}{A named \code{\link[MultiAssayExperiment]{Elist}} of
experiments compatible with the \code{MultiAssayExperiment} API}

\item{mPheno}{A \code{data.frame} of clinical data with patient identifiers
as rownames}
}
\value{
A named \code{list} of experiments
}
\description{
This function is intended to drop any unmatched samples from all of the
listed experiments that are not present in the rownames of the pheno data.
}
\author{
Marcel Ramos \email{mramos09@gmail.com}
}

