% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAenhanceClinical.R
\name{TCGAenhanceClinical}
\alias{TCGAenhanceClinical}
\title{Take a MultiAssayExperiment and include curated variables}
\usage{
TCGAenhanceClinical(MultiAssayExperiment = MultiAssayExperiment(), cancerCode,
  repoLocation = ".")
}
\arguments{
\item{MultiAssayExperiment}{A \linkS4class{MultiAssayExperiment} object}

\item{cancerCode}{A single string indicating the TCGA cancer code
(e.g., "PRAD")}

\item{repoLocation}{The directory location where the
"MultiAssayExperiment-TCGA" repository can be found}
}
\value{
A \code{\link{MultiAssayExperiment}} object
}
\description{
This function works on the \code{pData} of a
\code{\linkS4class{MultiAssayExperiment}} object to add curated variable
columns. It is recommended that the user run the scripts in the
\code{MultiAssayExperiment-TCGA} repository that build the "enhanced" type
of data. Please see the repository's README for more information.
}
