% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAprimaryTumors.R
\name{TCGAprimaryTumors}
\alias{TCGAprimaryTumors}
\title{Select primary tumors from TCGA datasets}
\usage{
TCGAprimaryTumors(multiassayexperiment)
}
\arguments{
\item{multiassayexperiment}{A \linkS4class{MultiAssayExperiment} with TCGA
data as obtained from \code{\link[curatedTCGAData]{curatedTCGAData}}}
}
\value{
A MultiAssayExperiment containing only primary tumor samples
}
\description{
Tumor selection is decided using the `sampleTypes` data. For 'LAML' datasets,
the primary tumor code used is "03" otherwise, "01" is used.
}
