% Generated by roxygen2: do not edit by hand
% Please edit documentation in:
% R/TCGAtranslateID.R
\name{TCGAtranslateID}
\alias{TCGAtranslateID}
\title{Translate TCGA Barcode to Universally Unique Identifiers and vice versa}
\usage{
TCGAtranslateID(identifier)
}
\arguments{
\item{identifier}{A \code{character} vector of either TCGA
barcodes or UUID identifiers}
}
\value{
A \code{data.frame} of original and translated identifiers
}
\description{
This function allows the user to enter a character vector of identifiers
to translate from barcodes to UUIDs and vice versa.
}
\author{
Marcel Ramos \email{mramos09@gmail.com}
}

