% Generated by roxygen2: do not edit by hand
% Please edit documentation in:
% R/TCGAtranslateUUID.R
\name{TCGAtranslateUUID}
\alias{TCGAtranslateUUID}
\title{Take UUIDs and return TCGA Barcodes}
\usage{
TCGAtranslateUUID(ids)
}
\arguments{
\item{ids}{A character vector of UUIDs}
}
\value{
A \code{character} vector of TCGA Barcode Identifiers
}
\description{
This function will use the GDC API to get metadata from Universally Unique
Identifiers (UUID)
}
\examples{
uuids <- c("0001801b-54b0-4551-8d7a-d66fb59429bf",
"002c67f2-ff52-4246-9d65-a3f69df6789e",
"003143c8-bbbf-46b9-a96f-f58530f4bb82")
TCGAtranslateUUID(uuids)

}
\author{
Marcel Ramos \email{mramos09@gmail.com}
}

