% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAutils-defunct.R
\name{splitAssays}
\alias{splitAssays}
\alias{makeSummarizedExperimentFromGISTIC}
\alias{TCGAutils-defunct}
\title{Defunct functions in TCGAutils}
\usage{
splitAssays(...)

makeSummarizedExperimentFromGISTIC(gistic, dataType, ...)
}
\arguments{
\item{...}{Additional arguments passed to 'RTCGAToolbox::getGISTICPeaks'.}

\item{gistic}{An `FirehoseGISTIC` object from `RTCGAToolbox`}

\item{dataType}{Either one of "ThresholdedByGene", "AllByGene", "Peaks"}
}
\value{
Defunct functions return errors
}
\description{
These functions have been moved to either `MultiAssayExperiment`
  or `RTCGAToolbox`
}
\details{
`splitAssays` was moved to `MultiAssayExperiment` and the TCGA
  specific implementation was named `TCGAsplitAssays` which resided in this
  package. `makeSummarizedExperimentFromGISTIC` now resides in
  `RTCGAToolbox`. See there for details.
}
\author{
L. Geistlinger, M. Ramos
}
