% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAutils-pkg.R
\docType{package}
\name{TCGAutils-package}
\alias{TCGAutils}
\alias{TCGAutils-package}
\title{TCGAutils: Helper functions for working with TCGA and MultiAssayExperiment
data}
\description{
TCGAutils is a toolbox to work with TCGA specific datasets. It allows the
user to manipulate and translate TCGA barcodes, conveniently convert a list
of data files to \link[GenomicRanges:GRangesList-class]{GRangesList}. Take
datasets from GISTIC and return a
\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}
class object. The package also provides functions for working with data from
the \code{curatedTCGAData}
experiment data package. It provides convenience functions for extracting
subtype metadata data and adding clinical data to existing
\link[MultiAssayExperiment:MultiAssayExperiment-class]{MultiAssayExperiment}
objects.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/waldronlab/TCGAutils/issues}
}

}
\author{
\strong{Maintainer}: Marcel Ramos \email{marcel.ramos@sph.cuny.edu} (\href{https://orcid.org/0000-0002-3242-0582}{ORCID})

Authors:
\itemize{
  \item Marcel Ramos \email{marcel.ramos@sph.cuny.edu} (\href{https://orcid.org/0000-0002-3242-0582}{ORCID})
  \item Lucas Schiffer
  \item Levi Waldron
}

Other contributors:
\itemize{
  \item Sean Davis [contributor]
  \item NCI (GrantNo.: U24CA289073) [funder]
}

}
