% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ID-translation.R
\name{UUID.barcode}
\alias{UUID.barcode}
\title{Translate study identifiers from barcode to UUID}
\usage{
UUID.barcode(file_ids, legacy = FALSE)
}
\arguments{
\item{identifier}{A \code{character} vector of identifiers}

\item{type}{A single \code{character} string indicating identifier type,
can use "file_id" (default) or "file_name" ignored if translating TCGA
barcodes}

\item{dataType}{A single \code{character} string indicating data type
used for searching (e.g., "Gene Expression Quantification")}
}
\value{
A \code{character} vector of alternate identifiers
}
\description{
This function allows the user to enter a character vector of identifiers
and use the GDC API to translate from TCGA barcodes to Universally Unique
Identifiers (UUID) and vice versa.
}
\examples{
## Translate UUIDs <--> TCGA Barcode
uuids <- c("0001801b-54b0-4551-8d7a-d66fb59429bf",
"002c67f2-ff52-4246-9d65-a3f69df6789e",
"003143c8-bbbf-46b9-a96f-f58530f4bb82")

UUID.barcode(uuids)

}
\author{
Sean Davis
}
