% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curatedTCGAData-helpers.R
\name{curatedTCGAData-helpers}
\alias{curatedTCGAData-helpers}
\alias{getSubtypeMap}
\alias{getClinicalNames}
\title{Helper functions for managing MultiAssayExperiment from
curatedTCGAData}
\usage{
getSubtypeMap(multiassayexperiment)

getClinicalNames(diseaseCode)
}
\arguments{
\item{multiassayexperiment}{A \linkS4class{MultiAssayExperiment} object}

\item{diseaseCode}{A TCGA cancer code (e.g., "BRCA")}
}
\value{
\itemize{
    \item{getSubtypeMap}: A \code{data.frame} with columns representing
    actual data variables and explanatory names
    \item{getClinicalNames}: A \code{vector} of names that correspond to
    a particular disease code.
}
}
\description{
Additional helper functions for cleaning and uncovering metadata
within a downloaded \code{MultiAssayExperiment} from \code{curatedTCGAData}.
The \code{getSubtypeMap} function provides a 2 column \code{data.frame}
with in-data variable names and an interpreted names. The
\code{getClinicalNames} function provides a vector of variable names that
exist in the \code{colData} slot of a downloaded \code{MultiAssayExperiment}
object. These variables are obtained from
\link[RTCGAToolbox]{getFirehoseData} by default and tend to be present
across most cancer codes.
}
\examples{
\dontrun{
library(curatedTCGAData)

coad <- curatedTCGAData(diseaseCode = "COAD",
    assays = "CNA*", dry.run = FALSE)
getSubtypeMap(coad)
}

getClinicalNames("COAD")

}
