% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diseaseCodes}
\alias{diseaseCodes}
\title{TCGA Cancer Disease Codes Table}
\format{
A data frame with 37 rows and 2 variables:
\describe{
\item{Study.Abbreviation}{Disease Code used in TCGA}
\item{Available}{Cancer datasets available via curatedTCGAData}
\item{SubtypeData}{Subtype curation data available via curatedTCGAData}
\item{Study.Name}{The full length study name (i.e., type of cancer)}
}
}
\source{
\url{https://gdc.cancer.gov/resources-tcga-users/tcga-code-tables/tcga-study-abbreviations}
}
\usage{
data("diseaseCodes")
}
\value{
The TCGA \code{diseaseCodes} table
}
\description{
A dataset for obtaining the cancer codes in TCGA for about 13 different
types of cancers.
}
\keyword{datasets}
