% Generated by roxygen2: do not edit by hand
% Please edit documentation in:
% R/getFileNames.R
\name{getFileNames}
\alias{getFileNames}
\title{Find the file names used in RTCGAToolbox}
\usage{
getFileNames(disease, runDate, CNA_SNP = FALSE, CNV_SNP = FALSE,
  CNA_Seq = FALSE, CNA_CGH = FALSE)
}
\arguments{
\item{disease}{The TCGA cancer disease code}

\item{runDate}{The single \code{string} used in the \code{getFirehoseData}
function}

\item{CNA_SNP}{A \code{logical} (default = FALSE) vector indicating whether to get the file
name from this data type}

\item{CNV_SNP}{A \code{logical} (default = FALSE) vector indicating whether to get the file
name from this data type}

\item{CNA_Seq}{A \code{logical} (default = FALSE) vector indicating whether to get the file
name from this data type}

\item{CNA_CGH}{A \code{logical} (default = FALSE) vector indicating whether to get the file
name from this data type}
}
\value{
A \code{character} vector of length one indicating the file name
}
\description{
Part of this function is from the RTCGAToolbox. It aims to extract the file
name used inside of the \link{getFirehoseData} function. The arguments of the
function parallel those in the \link{getFirehoseData} function. It is only
available for select data types.
}

