% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRangeNames.R
\name{getRangeNames}
\alias{getRangeNames}
\title{Obtain minimum necessary names for the creation of a GRangesList object}
\usage{
getRangeNames(namesVector)
}
\arguments{
\item{namesVector}{A \code{character} vector or names in a dataset}
}
\value{
The 3 to 4 length \code{character} vector of minimum necessary names to
create a \link{GRangesList} object
}
\description{
This function attempts to match chromosome, start position, end position and
strand names in the given character vector.
}
\examples{
myDataColNames <- c("txStart_position", "txEnd_position", "strand", "chromo",
"num_probes", "segment_mean")
getRangeNames(myDataColNames)

}

