% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplifyTCGA.R
\name{hidden-helpers}
\alias{hidden-helpers}
\alias{.makeListRanges}
\alias{.getRangesOfMir}
\alias{.getRangesOfSYMBOLS}
\alias{.getRangesOfCpG}
\title{Helper functions to get genomic ranges from different identifiers}
\usage{
.makeListRanges(x, gn)

.getRangesOfMir(x)

.getRangesOfSYMBOLS(x)

.getRangesOfCpG(x)
}
\arguments{
\item{x}{\code{character()} vector of either micro RNA identifiers
(\code{.getRangesOfMir}) or gene symbols (\code{.getRangesOfSYMBOLS}) or CpG probe
identifiers (\code{.getRangesOfCpG})}

\item{gn}{\code{GRanges()} with some of its names found in x or translated x}
}
\value{
list of length 2: "unmapped" \code{character()} symbols, "mapped" is a
\code{GRanges()} object with ranges of mapped symbols
}
\description{
Helper functions to get genomic ranges from different identifiers
}
\keyword{internal}
