% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplifyTCGA.R
\name{hidden-helpers}
\alias{hidden-helpers}
\alias{.makeListRanges}
\alias{.getRangesOfSYMBOLS}
\alias{.getRangesOfMir}
\title{A small document for helper functions}
\usage{
.makeListRanges(x, gn)

.getRangesOfSYMBOLS(x)

.getRangesOfMir(x)
}
\arguments{
\item{x}{A SummarizedExperiment containing hsa miR IDs as rownames}

\item{gn}{A GRanges object with some of its names found in x}
}
\value{
A list of length 2: unmapped (character vector) and mapped (GRanges)

list of length 2: "unmapped" is a character vector providing
unmapped symbols, "mapped" is a GRanges object with ranges of mapped symbols
}
\description{
A small document for helper functions
}
\keyword{internal}
