% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ID-translation.R
\name{ID-translation}
\alias{ID-translation}
\alias{UUIDtoBarcode}
\alias{UUIDtoUUID}
\alias{barcodeToUUID}
\alias{filenameToBarcode}
\title{Translate study identifiers from barcode to UUID and vice versa}
\usage{
UUIDtoBarcode(
  id_vector,
  from_type = c("case_id", "file_id", "aliquot_ids"),
  legacy = FALSE
)

UUIDtoUUID(id_vector, to_type = c("case_id", "file_id"), legacy = FALSE)

barcodeToUUID(barcodes, legacy = FALSE)

filenameToBarcode(filenames, legacy = FALSE)
}
\arguments{
\item{id_vector}{A \code{character} vector of UUIDs corresponding to
either files or cases (default assumes case_ids)}

\item{from_type}{Either \code{case_id} or \code{file_id} indicating the type of
\code{id_vector} entered (default "case_id")}

\item{legacy}{(logical default FALSE) whether to search the legacy archives}

\item{to_type}{The desired UUID type to obtain, can either be "case_id" or
"file_id"}

\item{barcodes}{A \code{character} vector of TCGA barcodes}

\item{filenames}{A \code{character} vector of filenames obtained from
the GenomicDataCommons}
}
\value{
A \code{data.frame} of TCGA barcode identifiers and UUIDs
}
\description{
These functions allow the user to enter a character vector of
identifiers and use the GDC API to translate from TCGA barcodes to
Universally Unique Identifiers (UUID) and vice versa. These relationships
are not one-to-one. Therefore, a \code{data.frame} is returned for all
inputs. The UUID to TCGA barcode translation only applies to file and case
UUIDs. Two-way UUID translation is available from 'file_id' to 'case_id'
and vice versa. Please double check any results before using these
features for analysis. Case / submitter identifiers are translated by
default, see the \code{from_type} argument for details. All identifiers are
converted to lower case.
}
\details{
Based on the file UUID supplied, the appropriate entity_id (TCGA barcode) is
returned. In previous versions of the package, the 'end_point' parameter
would require the user to specify what type of barcode needed. This is no
longer supported as `entity_id` returns the appropriate one.
}
\examples{
## Translate UUIDs >> TCGA Barcode

uuids <- c("0001801b-54b0-4551-8d7a-d66fb59429bf",
"002c67f2-ff52-4246-9d65-a3f69df6789e",
"003143c8-bbbf-46b9-a96f-f58530f4bb82")

UUIDtoBarcode(uuids, from_type = "file_id")

UUIDtoBarcode("ae55b2d3-62a1-419e-9f9a-5ddfac356db4", from_type = "case_id")

UUIDtoBarcode("d85d8a17-8aea-49d3-8a03-8f13141c163b", "aliquot_ids")

## Translate file UUIDs >> case UUIDs

uuids <- c("0001801b-54b0-4551-8d7a-d66fb59429bf",
"002c67f2-ff52-4246-9d65-a3f69df6789e",
"003143c8-bbbf-46b9-a96f-f58530f4bb82")

UUIDtoUUID(uuids)

## Translate TCGA Barcode >> UUIDs

fullBarcodes <- c("TCGA-B0-5117-11A-01D-1421-08",
"TCGA-B0-5094-11A-01D-1421-08",
"TCGA-E9-A295-10A-01D-A16D-09")

sample_ids <- TCGAbarcode(fullBarcodes, sample = TRUE)

barcodeToUUID(sample_ids)

participant_ids <- c("TCGA-CK-4948", "TCGA-D1-A17N",
"TCGA-4V-A9QX", "TCGA-4V-A9QM")

barcodeToUUID(participant_ids)

library(GenomicDataCommons)

fquery <- files() \%>\%
    filter(~ cases.project.project_id == "TCGA-COAD" &
        data_category == "Copy Number Variation" &
        data_type == "Copy Number Segment")

fnames <- results(fquery)$file_name[1:6]

filenameToBarcode(fnames)

}
\author{
Sean Davis, M. Ramos
}
