% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGRangesList.R
\name{makeGRangesList}
\alias{makeGRangesList}
\title{Convert raw TCGA Mutation data to GRangesList}
\usage{
makeGRangesList(inputData)
}
\arguments{
\item{inputData}{A \code{data.frame} or \code{list} class of TCGA
mutation data}
}
\value{
A \linkS4class{GRangesList} class object
}
\description{
This function takes the data.frame of raw data from the output of a TCGA
data pipeline and converts it to a \linkS4class{GRangesList} class.
Input data can be entered as either a \code{data.frame} with a sample
indicator titled as "tumor_sample_barcode" or "sample." If input data is a
entered as a list, all element names are expected to be TCGA
sample identifiers.
}
\author{
Marcel Ramos \email{mramos09@gmail.com}
}

