% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGRangesListFromExonFiles.R
\name{makeGRangesListFromExonFiles}
\alias{makeGRangesListFromExonFiles}
\title{Read Exon level files and create a GRangesList}
\usage{
makeGRangesListFromExonFiles(filepaths, sampleNames = NULL,
  rangesColumn = "exon")
}
\arguments{
\item{filepaths}{A vector of valid exon data file paths}

\item{sampleNames}{A vector of TCGA barcodes to be applied if not present in
the data}

\item{rangesColumn}{(default "exon") A single string indicating the name of the column in the
data containing the ranges information}
}
\value{
A \linkS4class{GRangesList} object
}
\description{
This function serves to read exon-level expression data. It works for exon
quantification (raw counts and RPKM) and junction quantification
(raw counts) files paths and represent such data as a
\linkS4class{GRangesList}. The data can be downloaded
via the TCGA Legacy Archive. File name and structure requirements are as
follows: The third position delimited by dots (".") in the file name should
be the universally unique identifier (UUID). The column containing the
ranged information is labeled "exon."
}
\examples{

pkgDir <- system.file("extdata", package = "TCGAutils", mustWork = TRUE)
exonFile <- list.files(pkgDir, pattern = "cation.txt$", full.names = TRUE)
makeGRangesListFromExonFiles(exonFile)

}
\author{
M. Ramos
}
