% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSummarizedExperimentFromGISTIC.R
\name{makeSummarizedExperimentFromGISTIC}
\alias{makeSummarizedExperimentFromGISTIC}
\title{Create a SummarizedExperiment from FireHose GISTIC}
\usage{
makeSummarizedExperimentFromGISTIC(gistic, dataType)
}
\arguments{
\item{gistic}{A \link[RTCGAToolbox]{FirehoseGISTIC-class} object}

\item{dataType}{Either one of "ThresholdedByGene", "AllByGene", "Peaks"}
}
\value{
A \code{SummarizedExperiment} object
}
\description{
Use the output of \code{getFirehoseData} to create a
\linkS4class{SummarizedExperiment}. This can be done for three types of
data, G-scores thresholded by gene, copy number by gene, and copy number by
peak regions.
}
\examples{

library(RTCGAToolbox)
co <- getFirehoseData("COAD", clinical = FALSE, GISTIC = TRUE,
    destdir = tempdir())
makeSummarizedExperimentFromGISTIC(co, "AllByGene")

}
\author{
L. Geistlinger, M. Ramos
}
