% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoPrintTCGA.R
\name{oncoPrintTCGA}
\alias{oncoPrintTCGA}
\title{OncoPrint for TCGA Mutation Assays}
\usage{
oncoPrintTCGA(multiassayexperiment, mutation = "*_Mutation-*",
  variantCol = "Variant_Classification", brewer_pal = "Set3")
}
\arguments{
\item{multiassayexperiment}{A MultiAssayExperiment preferably from
`curatedTCGAData``}

\item{mutation}{character(1) The name of the assay containing mutation data}

\item{variantCol}{character(1) The name of the metadata column containing
the mutation categories}

\item{brewer_pal}{character(1) The name of the `RColorBrewer` palette for
the different mutation types}
}
\description{
OncoPrint for TCGA Mutation Assays
}
\examples{

acc <- curatedTCGAData("ACC", "Mutation", FALSE)

oncoPrintTCGA(acc)

}
