% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplifyTCGA.R
\name{simplifyTCGA}
\alias{simplifyTCGA}
\title{Title All-in-one simplification of curatedTCGAData objects}
\usage{
simplifyTCGA(obj, keep = FALSE)
}
\arguments{
\item{obj}{A MultiAssayExperiment from curatedTCGAData}

\item{keep}{If FALSE (default), remove the original
MultiAssayExperiment elements that have simplified versions in the output.}
}
\value{
a MultiAssayExperiment with any gene expression, miRNA, copy number,
  and mutations converted to RangedSummarizedExperiment objects
}
\description{
Title All-in-one simplification of curatedTCGAData objects
}
\examples{
library(curatedTCGAData)

accmae <- curatedTCGAData("ACC",
    c("CNASNP", "Mutation", "miRNASeqGene", "GISTICT"),
    dry.run = FALSE)

simplifyTCGA(accmae)

}
\seealso{
mirToRanges, symbolsToRanges, qreduceTCGA
}
\author{
L. Waldron
}
