% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplifyTCGA.R
\name{simplifyTCGA}
\alias{simplifyTCGA}
\alias{symbolsToRanges}
\alias{mirToRanges}
\alias{CpGtoRanges}
\alias{qreduceTCGA}
\title{Functions to convert rows annotations to ranges and RaggedExperiment
to RangedSummarizedExperiment}
\usage{
simplifyTCGA(obj, keep.assay = FALSE, unmapped = TRUE)

symbolsToRanges(obj, keep.assay = FALSE, unmapped = TRUE)

mirToRanges(obj, keep.assay = FALSE, unmapped = TRUE)

CpGtoRanges(obj, keep.assay = FALSE, unmapped = TRUE)

qreduceTCGA(obj, keep.assay = FALSE, suffix = "_simplified")
}
\arguments{
\item{obj}{A MultiAssayExperiment object obtained from curatedTCGAData}

\item{keep.assay}{logical (default FALSE) Whether to keep the
\code{SummarizedExperiment} assays that have been converted to
\code{RangedSummarizedExperiment}}

\item{unmapped}{logical (default TRUE) Include an assay of data that was
not able to be mapped in reference database}

\item{suffix}{character (default "_simplified") A character string to append
to the newly modified assay for \code{qreduceTCGA}.}
}
\value{
A \linkS4class{MultiAssayExperiment} with any gene expression, miRNA,
copy number, and mutations converted to RangedSummarizedExperiment objects
}
\description{
This group of functions will convert row annotations as
either gene symbols or miRNA symbols to row ranges based on database
resources 'TxDB' and 'org.Hs' packages. It will also simplify the
representation of \linkS4class{RaggedExperiment} objects to
\linkS4class{RangedSummarizedExperiment}.
}
\details{
The original SummarizedExperiment containing either gene symbol
or miR annotations is replaced or supplemented by a
\linkS4class{RangedSummarizedExperiment} for those that could be mapped to
\linkS4class{GRanges}, and optionally another
\linkS4class{SummarizedExperiment} for annotations that
could not be mapped to \linkS4class{GRanges}.

RaggedExperiment mutation objects become a genes by patients
RangedSummarizedExperiment object containing '1' if there is a non-silent
mutation somewhere in the gene, and '0' otherwise as obtained from the
\code{Variant_Classification} column in the data.

"CNA" and "CNV" segmented copy number are reduced using a weighted mean in
the rare cases of overlapping (non-disjoint) copy number regions.

These functions rely on 'TxDb.Hsapiens.UCSC.hg19.knownGene' and
'org.Hs.eg.db' to map to the 'hg19' NCBI build. Users should use the
\code{liftOver} procedure for datasets that are provided against a different
reference genome (usually 'hg18'). An example of this procedure is provided
in the vignette.

\code{qreduceTCGA} will update \code{genome(x)} based on the NCBI reference annotation
which includes the patch number, e.g., GRCh37.p14, as provided by the
\code{seqlevelsStyle} setter, \code{seqlevelsStyle(gn) <- "NCBI"}. \code{qreduceTCGA}
uses the NCBI genome annotation as the default reference.
}
\examples{

library(curatedTCGAData)
library(GenomeInfoDb)

accmae <-
    curatedTCGAData(diseaseCode = "ACC",
    assays = c("CNASNP", "Mutation", "miRNASeqGene", "GISTICT"),
    version = "1.1.38",
    dry.run = FALSE)

## update genome annotation
rex <- accmae[["ACC_Mutation-20160128"]]

## Translate build to "hg19"
tgenome <- vapply(genome(rex), translateBuild, character(1L))
genome(rex) <- tgenome

accmae[["ACC_Mutation-20160128"]] <- rex

simplifyTCGA(accmae)

}
\author{
L. Waldron
}
