% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga.R
\name{tcga}
\alias{tcga}
\title{Set TCGA data parameters}
\usage{
tcga(primary = NULL, standard = TRUE, rangeID = character(),
  idFUN = NULL)
}
\arguments{
\item{primary}{The column name for primary samples/specimen, defaults to
"Tumor_Sample_Barcode" (case ignored).}

\item{standard}{logical (default TRUE) whether to use standard range column
indicators found in the tcga data, these include "chromosome",
"start_position", "end_position", and "strand"}

\item{rangeID}{An optional \code{character} vector of length four indicating
genomic positions when standard argument is FALSE}

\item{idFUN}{A function that helps in the parsing of barcode identifiers
(barcode used by default)}
}
\value{
A \code{list} of arguments for use by \link{makeGRangesList}
}
\description{
This function is for use inside the \link{makeGRangesList} function for
converting raw data. The parameters in this function indicate properties
of the TCGA data.
}
\author{
Marcel Ramos \email{mramos09@gmail.com}
}

