% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAbarcode.R
\name{TCGAbarcode}
\alias{TCGAbarcode}
\title{Parse data from TCGA barcode}
\usage{
TCGAbarcode(barcodes, participant = TRUE, sample = FALSE, portion = FALSE,
  plate = FALSE, center = FALSE, index = NULL)
}
\arguments{
\item{barcodes}{A character vector of TCGA barcodes}

\item{participant}{Logical (default TRUE) participant identifier chunk}

\item{sample}{Logical (default FALSE) includes the numeric sample code of
the barcode}

\item{portion}{Logical (default FALSE) includes the portion and analyte
codes of the barcode}

\item{plate}{Logical (default FALSE) returns the plate value}

\item{center}{Logical (default FALSE) returns a matrix with the plate and
center codes}

\item{index}{A numerical vector of TCGA barcode positions desired}
}
\value{
A character vector or data matrix of TCGA barcode information
}
\description{
This function returns the specified snippet of information obtained from
the TCGA barcode.
}
\examples{
library(TCGAbiolinks)
query <- TCGAquery(tumor = "OV", level = 3)
barcodes <- unlist(strsplit(query$barcode[1], ","))
TCGAsamplebc <- TCGAbarcode(barcodes)

}
\author{
Marcel Ramos \email{mramos09@gmail.com}
}

