% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAcleanExpList.R
\name{TCGAcleanExpList}
\alias{TCGAcleanExpList}
\title{Remove unmatched samples from a list of experiments}
\usage{
TCGAcleanExpList(experiments, colData)
}
\arguments{
\item{experiments}{A named \link[MultiAssayExperiment]{ExperimentList}
of experiments compatible with the \code{MultiAssayExperiment} API}

\item{colData}{A \code{data.frame} of clinical data with patient identifiers
as rownames}
}
\value{
A named \code{list} or \link[MultiAssayExperiment]{ExperimentList}
if the \code{MultiAssayExperiment} package is available
}
\description{
This function is intended to drop any unmatched samples from all of the
listed experiments that are not present in the rownames of the pheno data.
}
\author{
Marcel Ramos \email{mramos09@gmail.com}
}
