% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAextract.R
\name{TCGAextract}
\alias{TCGAextract}
\title{Extract data from \code{FirehoseData} object into \code{ExpressionSet} or
\code{GRangesList} object}
\usage{
TCGAextract(object, type = c("Clinical", "RNASeqGene", "miRNASeqGene",
  "RNASeq2GeneNorm", "CNASNP", "CNVSNP", "CNASeq", "CNACGH", "Methylation",
  "Mutation", "mRNAArray", "miRNAArray", "RPPAArray", "GISTIC", "GISTICA",
  "GISTICT"), ...)
}
\arguments{
\item{object}{A \code{FirehoseData} object from which to extract data.}

\item{type}{The type of data to extract from the "FirehoseData" object,
see type section.}
}
\value{
Either an \linkS4class{ExpressionSet} object or a
\linkS4class{GRangesList} object.
}
\description{
This function processes data from a \code{\linkS4class{FirehoseData}}
object from the \code{RTCGAToolbox} package. Raw data is converted to
conventional Bioconductor objects. The function returns an
\linkS4class{ExpressionSet} or \linkS4class{GRangesList} class object. Note:
this function works best with the modifications found in the github fork:
\code{LiNk-NY/RTCGAToolbox}. In cases where range data are found
(i.e., "mutations") the default extraction method is used
(see makeGRangesListFromTCGA).
}
\section{type}{

Choices include: "RNAseqGene", "Clinical", "miRNASeqGene",
"RNASeq2GeneNorm", "CNASNP", "CNVSNP", "CNASeq", "CNACGH", "Methylation",
"Mutation", "mRNAArray", "miRNAArray", "RPPAArray", "GISTIC", "GISTICA",
"GISTICT". The "GISTICA" type of dataset represents GISTIC data by all
genes. "GISTICT" represents data thresholded by genes. To get both types
in a list, use "GISTIC". Lowercase entries are also valid inputs.
}

\examples{
\dontrun{
library(RTCGAToolbox)
dataFolder <- normalizePath("~/Documents/data")
coadmut <- getFirehoseData("COAD", runDate = "20151101", Mutation = TRUE,
                         destdir = dataFolder)
cm <- TCGAextract(coadmut, "Mutation")
}
}
\author{
Marcel Ramos \email{marcel.ramos@roswellpark.org}
}
