% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAtranslateID.R
\name{TCGAtranslateID}
\alias{TCGAtranslateID}
\title{Translate study identifiers from barcode to UUID and vice versa}
\usage{
TCGAtranslateID(identifier, type = "file_id", dataType = NULL)
}
\arguments{
\item{identifier}{A \code{character} vector of identifiers}

\item{type}{A single \code{character} string indicating identifier type,
can use "file_id" (default) or "file_name" ignored if translating TCGA
barcodes}

\item{dataType}{A single \code{character} string indicating data type
used for searching (e.g., "Gene Expression Quantification")}
}
\value{
A \code{character} vector of alternate identifiers
}
\description{
This function allows the user to enter a character vector of identifiers
and use the GDC API to translate from TCGA barcodes to Universally Unique
Identifiers (UUID) and vice versa.
}
\examples{
## Translate UUIDs <--> TCGA Barcode
uuids <- c("0001801b-54b0-4551-8d7a-d66fb59429bf",
"002c67f2-ff52-4246-9d65-a3f69df6789e",
"003143c8-bbbf-46b9-a96f-f58530f4bb82")

TCGAtranslateID(uuids)

## Translate TCGA Barcode <--> UUIDs
barcodes <- c("TCGA-B0-5117-11A-01D-1421-08",
"TCGA-B0-5094-11A-01D-1421-08",
"TCGA-E9-A295-10A-01D-A16D-09")
pt_identifiers <- TCGAbarcode(barcodes)
TCGAtranslateID(pt_identifiers, type="file_id")

## Complex Example
exampleCodes <- c("TCGA-CK-4948",
"TCGA-D1-A17N",
"TCGA-4V-A9QX",
"TCGA-4V-A9QM")
exampleType <- "Gene Expression Quantification"
TCGAtranslateID(exampleCodes, dataType = exampleType)

}
\author{
Marcel Ramos \email{mramos09@gmail.com}
}
