% Generated by roxygen2: do not edit by hand
% Please edit documentation in:
% R/TCGAtranslateUUID.R
\name{TCGAtranslateUUID}
\alias{TCGAtranslateUUID}
\title{Take UUIDs and return TCGA Barcodes}
\usage{
TCGAtranslateUUID(ids)
}
\arguments{
\item{ids}{A character vector of UUIDs}
}
\value{
A \code{character} vector of TCGA Barcode Identifiers
}
\description{
This function will use the GDC API to get metadata from Universally Unique
Identifiers (UUID)
}
\examples{
uuids <- c("0004d251-3f70-4395-b175-c94c2f5b1b81",
"000d566c-96c7-4f1c-b36e-fa2222467983",
"0011a67b-1ba9-4a32-a6b8-7850759a38cf")
TCGAtranslateUUID(uuids)

}
\author{
MRamos
}

