% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAutils-defunct.R
\name{splitAssays}
\alias{splitAssays}
\alias{makeSummarizedExperimentFromGISTIC}
\alias{TCGAutils-defunct}
\title{Defunct functions in TCGAutils}
\usage{
splitAssays(...)

makeSummarizedExperimentFromGISTIC(gistic, dataType, ...)
}
\arguments{
\item{...}{Additional arguments passed to 'RTCGAToolbox::getGISTICPeaks'.}

\item{gistic}{An \code{FirehoseGISTIC} object from \code{RTCGAToolbox}}

\item{dataType}{Either one of "ThresholdedByGene", "AllByGene", "Peaks"}
}
\value{
Defunct functions return errors
}
\description{
These functions have been moved to either \code{MultiAssayExperiment}
or \code{RTCGAToolbox}
}
\details{
\code{splitAssays} was moved to \code{MultiAssayExperiment} and the TCGA
specific implementation was named \code{TCGAsplitAssays} which resided in this
package. \code{makeSummarizedExperimentFromGISTIC} now resides in
\code{RTCGAToolbox}. See there for details.
}
\author{
L. Geistlinger, M. Ramos
}
