% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAutils-pkg.R
\docType{package}
\name{TCGAutils-package}
\alias{TCGAutils}
\alias{TCGAutils-package}
\title{TCGAutils: Helper functions for working with TCGA and MultiAssayExperiment
data}
\description{
TCGAutils is a toolbox to work with TCGA specific datasets. It allows the
user to manipulate and translate TCGA barcodes, conveniently convert a list
of data files to \linkS4class{GRangesList}. Take datasets from GISTIC and
return a \linkS4class{SummarizedExperiment} class object. The package also
provides functions for working with data from the \code{curatedTCGAData}
experiment data package. It provides convenience functions for extracting
subtype metadata data and adding clinical data to existing
\linkS4class{MultiAssayExperiment} objects.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/waldronlab/TCGAutils/issues}
}

}
\author{
\strong{Maintainer}: Marcel Ramos \email{marcel.ramos@roswellpark.org}

Authors:
\itemize{
  \item Lucas Schiffer
  \item Levi Waldron
}

Other contributors:
\itemize{
  \item Sean Davis [contributor]
}

}
